% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggnet.R
\name{coord_place}
\alias{coord_place}
\title{Set coordinates of nodes in network graph}
\usage{
coord_place(network, layout = "fruchtermanreingold", plot.layout = NULL)
}
\arguments{
\item{network}{This is a network-class object that is being plotted.}

\item{layout}{Identifies the layout algorithm to use. Default is 
Fruchterman-Reingold.}

\item{plot.layout}{Needed for certain layout types. Default is NULL. You 
can ignore.}
}
\description{
Set coordinates of nodes using a variety of force-directed layout
algorithms. Can be used within the main graphing function to create
a random layout of nodes each time or be run independently to 
ensure a consistent layout of nodes within a network during visualization.
}
\examples{
data(example_network)
coords <- coord_place(net)
}
\keyword{ggplot2}
\keyword{network,}
\keyword{visualization,}

