% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggnet.R
\name{ggnet}
\alias{ggnet}
\title{Graph social network using ggplot2}
\usage{
ggnet(net, direct = FALSE, color = NULL, names = NULL, shape = NULL,
  size = NULL, edge.val = NULL, title = NULL, legend = FALSE,
  coords = NULL, layout = "fruchtermanreingold", plot.layout = NULL,
  palette = "Set1", gradient = FALSE, alpha = 1)
}
\arguments{
\item{net}{This is a network-class object that is being plotted.}

\item{direct}{Boolean value identifying if the network is directed or not. 
Default is FALSE.}

\item{color}{Identifies the attribute (as a string) to color nodes by. 
Default is NULL.}

\item{names}{Identifies the attribute (as a string) to name nodes by. 
Default is NULL.}

\item{shape}{Identifies the attribute (as a string) to shape nodes by. 
Default is NULL.}

\item{size}{Identifies the attribute (as a string) to size nodes by. 
Default is NULL.}

\item{edge.val}{A numeric matrix to make edges differentially transparent 
based on value. Default is NULL.}

\item{title}{A string containing title of visualization. Default is NULL.}

\item{legend}{Boolean value to indicate if a legend should be printed if 
attributes are used. Default is FALSE.}

\item{coords}{An optional specification of coordinates (from coord_place) 
for network. Default is NULL, which will produce a random layout of nodes.}

\item{layout}{Specifies layout algorithm for nodes. Default is the 
Fruchterman-Reingold algorithm.}

\item{plot.layout}{Needed for certain layout types. Default is NULL. You 
can ignore.}

\item{palette}{Which ggplot2 color palette to use for plotting. Default 
is 'Set1'.}

\item{gradient}{Boolean value indicating if the 'color' attribute should be
a gradient or distinct color set. Default is FALSE.}

\item{alpha}{Numeric value from 0 to 1 to optionally change the transparency 
of color in the graph. Default is 1.}
}
\description{
Plot a network visualization using ggplot2 as the underlying architecture. 
The function can utilize a number of attributes for nodes and edges, can 
plot directed/undirected graphs, and utilize a variety of color schemes.
}
\examples{
data(example_network)
plot <- ggnet(net, direct = TRUE, color = 'age', title = 'Network graph', 
              names = 'vertex.names', gradient = TRUE)
}
\keyword{ggplot2}
\keyword{network,}
\keyword{visualization,}

