% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{require_pkgs}
\alias{require_pkgs}
\title{Require packages}
\usage{
require_pkgs(pkgs)
}
\arguments{
\item{pkgs}{vector of character values}
}
\description{
This function is taken from the \code{GGally} package.
}
\author{
Barret Schloerke
}
\keyword{internal}

